Added EJS server-side prerendering
I've been experiencing some problems with high CLS values. This shows up
in the Google Search Console, and is likely causing SEO issues. The main
issue is likely that I have the website render the details from an array
on the client side, since it's a static page. This has some downsides:

1. The client doesn't get the full page until the JavaScript finishes,
which can cause layout shifts.
2. Users with JavaScript blocked will see... nothing! (Or at least a
very bare page.)

To remedy this, I'm now taking advantage of Netlify's automatic build
service to pre-render this data. The index is now written with EJS, and
any data that can be pre-filled from anywhere is. The detail data was
also moved out into a separate JSON file for ease of updating.

I should have been doing something like this for a long time, but
I'm not really a web developer so I'm still stuck in the past with this stuff.