fix: NixOS usage instructions
Before, the documentation says to install CraftOS-PC before running it,
which is not necessary with the `nix-shell -p` incantation.

The documentation also says that these instructions are for NixOS,
whereas they should work with any machine that has Nix (for example,I
have a Linux Mint install that has Nix installed as well).